// give-away-vite/give-away/src/api/api.ts
export interface AuctionListResponse {
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'auctionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'startDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'endDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof AuctionListResponse
     */
    'depositFee'?: number;
    /**
     * 
     * @type {AuctionStatus}
     * @memberof AuctionListResponse
     */
    'status'?: AuctionStatus;
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'shopId'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuctionListResponse
     */
    'auctionItemId'?: string;
    /**
     * 
     * @type {ShopDetailResponse}
     * @memberof AuctionListResponse
     */
    'shop'?: ShopDetailResponse;
}
// ...
export const AuctionStatus = {
    Pending: 'Pending',
    Rejected: 'Rejected',
    Approved: 'Approved',
    OnGoing: 'OnGoing',
    Finished: 'Finished'
} as const;

export type AuctionStatus = typeof AuctionStatus[keyof typeof AuctionStatus];
// ...
export interface ShopDetailResponse {
    /**
     * 
     * @type {string}
     * @memberof ShopDetailResponse
     */
    'shopId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ShopDetailResponse
     */
    'address'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ShopDetailResponse
     */
    'staffId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ShopDetailResponse
     */
    'phone'?: string | null;
}